Fixed crash when combining evidence
Evidence combination conversations have always been treated strangely
compared to other conversations, in that the encounter they're placed in
is temporary holder created and then deleted by the current location,
such that an evidence combination conversation can potentially have
several parent encounters.  Conversation::Reset() performs an operation
on pState, which may in this circumstance be a leftover from a
now-deleted encounter, which leads to a crash.

To fix this, I've done a few things that should prevent this from
happening in the future:

- Rather than having the location object create the encounter, a new
encounter is now created for and permanently associated with each
evidence combination conversation, and these encounters are directly
handed to the location.

- pState->ResetForConversation() is no longer called in
Conversation::Reset(); instead, it's split into a new ResetState()
method that is called where state.ResetForConversation() used to be
called in Encounter.

- Also, for good measure (though it ended up not being necessary),
ppItems is set to NULL in ~EvidenceSelector().