Case creator staging #2
Checking in some more changes to already-existing files that are needed
for the case creator.

Also removing everything surrounding BeginCheckPartnerAction et al. from
the code base.  That was never used in case 1, as it was replaced by the
far superior and more versatile partner-present unlock condition.  Don't
really want people to be able to access this inferior way of doing
things when they start making cases.