XML formatting versioning
- Added the ability to specify a formatting version in XML, which can be
used by XmlReader to know what version of an XML document to expect.
This can be used to make it so that types are backwards-compatible with
the XML format in the case file that already was shipped, while also
being able to use the new automated XML save/load system.

- Fixed a bug where subclasses of classes derived from XmlStorableObject
overwrite the elements in the storage handler of its parent class when
it goes to add its own elements.

- Added a slight improvement where the XmlStorableObject now detects
when an element has no child elements, and simply writes an empty
element in that case instead of starting and then immediately ending an
element on different lines.

- Finally, added the ability to verify in debug mode that when we start
and then end an element, that we're on the same element we started when
we proceed to end that element.  This can be used to immediately catch
otherwise nasty XML bugs where, for example, we accidentally call
EndElement() twice in custom-written XML when we meant to only call it
once.