Localization improvements
- Added support for a file containing localized content that is separate
from the file containing non-localizable content.

- Added interfaces that allow controls that read from that file to
automatically refresh themselves when the current language changes.

- Added a screen in which users can select a new language based on the
localized resource files installed.

- Updated buttons to make it so that the x-position can be specified in
terms of where either the center or the right of the control should be,
in order to make proper layout after language selection easier.

- Made the selection highlight in selectors draw as a nine-grid image
instead of as a single image, in order to accommodate selectors of
different sizes.

- Improved UTF-8 handling to no longer rely on exception catching, since
exceptions are expensive to work with and confused the Code::Blocks
debugger.

- Moved some still-hardcoded strings to the file containing localized
content.

- Created a shared utils set of files to handle things like parsing
dialog strings, in order to make it so that the game executable and the
case creator can share text formatting logic to ensure a consistent
experience between the two.

- Renamed the localizable content object to "gpLocalizableContent",
instead of the inconsistently named "pgLocalizableContent".