Updater improvements and localization improvements
- Completely overhauling the way the updater works: now, instead of
doing file replacement in-process, it instead outputs a batch file/bash
script that is then run after we finish downloading and staging
everything.  This will make it so that we can update *everything*,
including SDL and the updater itself, instead of only things relating to
the MLI executable.

- Moved a bunch more strings to the localization file, including (I
think) all of the strings in the updater.  The updater is finally
localizable!  Yay!

- Shifted the alignment of some buttons to work when localized.

- Changed sprintf to snprintf to avoid buffer overruns.